requirements.txt
torch==1.12.1+cu113 torchvision==0.13.1+cu113 torchaudio==0.12.1 --extra-index-url https://download.pytorch.org/whl/cu113
pandas==2.2.1
tqdm==4.66.2
dlib==19.24.2
opencv-python==4.9.0.80
efficientnet-pytorch==0.7.1
wandb==0.16.4
scikit-learn==1.4.1.post1
albumentations==1.4.2